Moved get_access_token from OrgConfig to SfdxOrgConfig Now throws SfdxOrgException rather than Exception Added sfdx property to SfdxOrgConfig to aid writing mocks; maybe there's a better way?
